Move calculated question update logic to instance method
