+++ packages/lan/app/routes/apiv1/surveyResponseAnswer.js
@@ -3,8 +3,6 @@ import asyncHandler from 'express-async-handler';
 import { Op } from 'sequelize';
 import { NotFoundError, InvalidOperationError, InvalidParameterError } from 'shared/errors';
 import { SURVEY_TYPES, PROGRAM_DATA_ELEMENT_TYPES } from 'shared/constants';
-import { runCalculations } from 'shared/utils/calculations';
-import { getStringValue } from 'shared/utils/fields';
 
 export const surveyResponseAnswer = express.Router();
 
@@ -12,14 +10,7 @@ surveyResponseAnswer.put(
   '/vital/:id',
   asyncHandler(async (req, res) => {
     const { db, models, user, params, getLocalisation } = req;
-    const {
-      SurveyResponseAnswer,
-      SurveyResponse,
-      Survey,
-      VitalLog,
-      ProgramDataElement,
-      SurveyScreenComponent,
-    } = models;
+    const { SurveyResponseAnswer, SurveyResponse, Survey, VitalLog, ProgramDataElement } = models;
     const { id } = params;
 
     const localisation = await getLocalisation();
@@ -55,11 +46,6 @@ surveyResponseAnswer.put(
       throw new InvalidParameterError('New value is the same as previous value.');
     }
 
-    const screenComponents = await SurveyScreenComponent.getComponentsForSurvey(
-      answerObject.surveyResponse.survey.id,
-    );
-    const calculatedScreenComponents = screenComponents.filter(c => c.calculation);
-
     await db.transaction(async () => {
       const { newValue = '', reasonForChange, date } = req.body;
       await VitalLog.create({
@@ -71,55 +57,7 @@ surveyResponseAnswer.put(
         answerId: id,
       });
       await answerObject.update({ body: newValue });
-
-      // Check if any calculated question is affected by this change
-      let answers;
-      let calculatedValues;
-      for (const component of calculatedScreenComponents) {
-        if (component.calculation.includes(answerObject.ProgramDataElement.code) === false) {
-          continue;
-        }
-
-        // Grab answers only once
-        if (!calculatedValues) {
-          const response = await SurveyResponse.findByPk(answerObject.surveyResponse.id, {
-            include: [
-              {
-                required: true,
-                model: SurveyResponseAnswer,
-                as: 'answers',
-              },
-            ],
-          });
-          answers = response.answers;
-          const values = {};
-          answers.forEach(answer => {
-            values[answer.dataElementId] = answer.body;
-          });
-          calculatedValues = runCalculations(screenComponents, values);
-        }
-
-        // Modify survey response answer for calculated value and create log
-        const calculatedAnswer = answers.find(
-          answer => answer.dataElementId === component.dataElement.id,
-        );
-        if (!calculatedAnswer) continue;
-
-        const stringValue = getStringValue(
-          component.dataElement.type,
-          calculatedValues[component.dataElement.id],
-        );
-        const newCalculatedValue = stringValue ?? '';
-        await VitalLog.create({
-          date,
-          reasonForChange,
-          previousValue: calculatedAnswer.body,
-          newValue: newCalculatedValue,
-          recordedById: user.id,
-          answerId: calculatedAnswer.id,
-        });
-        await calculatedAnswer.update({ body: newCalculatedValue });
-      }
+      await answerObject.upsertCalculatedQuestions({ date, reasonForChange, user });
     });
 
     res.send(answerObject);
